refactor(NoValidMailboxes): Move file to their own package
