+++ app/src/main/java/com/infomaniak/mail/ui/noValidMailboxes/NoValidMailboxesFragment.kt
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.infomaniak.mail.ui.login
+package com.infomaniak.mail.ui.noValidMailboxes
 
 import android.os.Bundle
 import android.view.LayoutInflater
